+++ lib/client/sse_client.dart
@@ -56,10 +56,10 @@ class SseClient extends StreamChannelMixin<String> {
     });
   }
 
-  @Deprecated('Use onConnect instead.')
+  @Deprecated(
+      'Outgoing messages are now buffered until a connection is established.'
+      'This should no longer be required and will be removed.')
   Stream<Event> get onOpen => _eventSource.onOpen;
-
-  Future<void> get onConnect => _eventSource.onOpen.first;
 
   /// Add messages to this [StreamSink] to send them to the server.
   ///